fix: Switch cmdLogger to use composition instead of embedding
Tracing through the code execution indicated that the
`cmdLogger.Write()` method was not actually being called but that
`bytes.Buffer.Write()` was being called instead. This results in
`cmdLogger.writeLines()` not running until the command has finished and
the remaining buffer output is flushed in one long string full of `\n`
and `\r\n` sequences.

Changed the `cmdLogger` struct from using embedding to composition to
ensure that the correct `Write()` method is called.

Changelog: Fix printing of update module stdout in real-time
Ticket: None
Signed-off-by: Domenic Rodriguez <domenic.rodriguez@geckorobotics.com>